DL-Central Nervous defficiencies detection
This is just the part of the project I did. I can't provide all the project, nor the data. 

But the idea of the project was the identify the type of central nervous defficiencies of an individu suffering from Alzheimer's disease (dementia). Multiple tests were carried out in the laboratory on small lab animals (rats). They were exposed to radiations in different levels to simulate the Alzheimer's disease. And a whole set up was built to collected useful data (what data was collected ? What about the preprocessing? what type of models did we built? If you interested in the project, ping me by email)